Draft: split DrawingView from Draft.py
